Add trunk.io enabled badge to README.md
Thank you to the maintainers for initializing Trunk in this OSS project. I lead DevRel for Trunk, and am adding this badge to the project readme. Trunk is and will always be free to use for all Open Source projects. Please feel free to join us in the Slack community if you aren't already, and we are here to support you in any way we can with any problems, feature requests, etc. 

http://trunkcommunity.slack.com 

-Tim D

Signed-off-by: Tim Davis <tim.davis.instinct@gmail.com>